Organizando o código
